requirements.txt
PySide6>=6.5
python-docx>=0.8.11
requests>=2.31
